# RocketTE(WebEdition)/bzwapps/app/bzshared/doc/swagger/bzshared.yaml
swagger: "2.0"
info:
  description: "API document for Rocket BlueZone Web"
  version: "1.0.0"
  title: "Rocket BlueZone Shared APIs"
  contact:
    email: "jgao@rocketsoftware.com"
  license:
    name: "Rocket Software Commercial License"
host: "rs28:30004"
basePath: "/ZLUX/plugins/com.rs.bzshared/services"
...
schemes:
- "https"
- "http"
paths:
...
  /script3270:
    get:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "script3270"
      operationId: "getScript3270"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
  /script5250:
    get:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "script5250"
      operationId: "getScript5250"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
...
  /script:
    put:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "script"
      operationId: "createScript"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"      
...